requirements.txt
psycopg[binary]
Pillow
Flask
